# Conflicts:
#	app/Http/Controllers/Cooperation/Tool/HighEfficiencyBoilerController.php
#	app/Http/Controllers/Cooperation/Tool/RoofInsulationController.php
#	app/Models/BuildingCoachStatus.php
#	database/seeds/DatabaseSeeder.php
#	public/css/app.css
#	resources/lang/nl/woningdossier.php
#	routes/web.php

+++ app/Services/UserService.php
@@ -12,6 +12,7 @@ class UserService
 {
     public static function deleteUser(User $user)
     {
+
         // if the user is only associated with 1 cooperation, we delete the whole account with all its data
         // else we delete the relation between the cooperation and the user.
         if ($user->cooperations()->count() === 1) {
@@ -33,10 +34,13 @@ class UserService
             $user->energyHabit()->withoutGlobalScope(GetValueScope::class)->delete();
             // remove the motivations from a user
             $user->motivations()->delete();
+            // remove the notification settings
+            $user->notificationSettings()->delete();
             // remove the progress from a user
             //$user->progress()->delete();
             // delete the cooperation from the user, belongsToMany so no deleting here.
             $user->cooperations()->detach();
+
 
             // finally remove the user itself :(
             $user->delete();
@@ -45,10 +49,13 @@ class UserService
             $currentCooperation = Cooperation::find(HoomdossierSession::getCooperation());
             $user->cooperations()->detach($currentCooperation);
 
-            // the user still exists, so we have to logout the user
-            HoomdossierSession::destroy();
-            \Auth::logout();
-            request()->session()->invalidate();
+            // we only want to do this if the user is deleting himself. Otherwise admins would randomly logout.
+            if (\Auth::id() == $user->id) {
+                // the user still exists, so we have to logout the user
+                HoomdossierSession::destroy();
+                \Auth::logout();
+                request()->session()->invalidate();
+            }
         }
     }
 }
